+++ Source/dom/index.js
@@ -1,23 +1,23 @@
 import { exportObject } from './export'
 
-import { Rectangle } from './Rectangle'
+import { Document } from './models/Document'
+import { Library } from './models/Library'
+import { SharedStyle } from './models/SharedStyle'
+import { Rectangle } from './models/Rectangle'
+import { AnimationType, BackTarget } from './models/Flow'
 
 import { Style } from './style/Style'
 
-import { Document } from './components/Document'
-import { Group } from './components/Group'
-import { Text } from './components/Text'
-import { Image } from './components/Image'
-import { Shape } from './components/Shape'
-import { Artboard } from './components/Artboard'
-import { Page } from './components/Page'
-import { SymbolMaster } from './components/SymbolMaster'
-import { SymbolInstance } from './components/SymbolInstance'
-import { Library } from './components/Library'
-import { HotSpot } from './components/HotSpot'
-import { SharedStyle } from './components/SharedStyle'
+import { Group } from './layers/Group'
+import { Text } from './layers/Text'
+import { Image } from './layers/Image'
+import { Shape } from './layers/Shape'
+import { Artboard } from './layers/Artboard'
+import { Page } from './layers/Page'
+import { SymbolMaster } from './layers/SymbolMaster'
+import { SymbolInstance } from './layers/SymbolInstance'
+import { HotSpot } from './layers/HotSpot'
 
-import { AnimationType, BackTarget } from './components/Flow'
 import { Types } from './enums'
 import { wrapObject } from './wrapNativeObject'
