+++ Source/dom/layers/Artboard.js
@@ -1,6 +1,6 @@
 import { DefinedPropertiesKey } from '../WrappedObject'
 import { Group } from './Group'
-import { Rectangle } from '../Rectangle'
+import { Rectangle } from '../models/Rectangle'
 import { Types } from '../enums'
 import { Factory } from '../Factory'
