+++ Source/dom/layers/HotSpot.js
@@ -1,6 +1,6 @@
 import { DefinedPropertiesKey } from '../WrappedObject'
 import { Layer } from './Layer'
-import { Rectangle } from '../Rectangle'
+import { Rectangle } from '../models/Rectangle'
 import { Types } from '../enums'
 import { Factory } from '../Factory'
 import { wrapObject } from '../wrapNativeObject'
