+++ Source/dom/layers/Image.js
@@ -1,7 +1,7 @@
 import { DefinedPropertiesKey } from '../WrappedObject'
 import { StyledLayer } from './StyledLayer'
-import { ImageData } from './ImageData'
-import { Rectangle } from '../Rectangle'
+import { ImageData } from '../models/ImageData'
+import { Rectangle } from '../models/Rectangle'
 import { Types } from '../enums'
 import { Factory } from '../Factory'
 import { wrapObject } from '../wrapNativeObject'
