+++ Source/dom/layers/Shape.js
@@ -2,7 +2,7 @@ import { DefinedPropertiesKey } from '../WrappedObject'
 import { StyledLayer } from './StyledLayer'
 import { Types } from '../enums'
 import { Factory } from '../Factory'
-import { Rectangle } from '../Rectangle'
+import { Rectangle } from '../models/Rectangle'
 
 // TODO: set and modify path
