+++ Source/dom/layers/StyledLayer.js
