+++ Source/dom/layers/Text.js
@@ -1,6 +1,6 @@
 import { DefinedPropertiesKey } from '../WrappedObject'
 import { StyledLayer } from './StyledLayer'
-import { Rectangle } from '../Rectangle'
+import { Rectangle } from '../models/Rectangle'
 import { Types } from '../enums'
 import { Factory } from '../Factory'
