+++ Source/dom/layers/__tests__/HotSpot.test.js
