+++ Source/dom/models/ImageData.js
