+++ Source/dom/models/SharedStyle.js
