+++ Source/dom/models/__tests__/SharedStyle.test.js
@@ -1,7 +1,7 @@
 /* globals expect, test */
 /* eslint-disable no-param-reassign */
 
-import { Text, Shape } from '../../'
+import { Text, Shape } from '../..'
 import { createSharedStyle } from '../../../test-utils'
 
 test(
