+++ Source/dom/style/Gradient.js
@@ -1,7 +1,7 @@
 import { WrappedObject, DefinedPropertiesKey } from '../WrappedObject'
 import { toArray, isNativeObject } from '../utils'
 import { GradientStop } from './GradientStop'
-import { Point } from '../Point'
+import { Point } from '../models/Point'
 import { Types } from '../enums'
 
 const GradientTypeMap = {
